+++ easyeasm/main.go
@@ -1,9 +1,12 @@
 package main
 
 import (
+	"database/sql"
 	"fmt"
+	"log"
 	"os"
 	"strings"
+	"time"
 
 	"github.com/sethlaw/EasyEASM/pkg/active"
 	"github.com/sethlaw/EasyEASM/pkg/configparser"
@@ -22,6 +25,16 @@ func main() {
 	// parse the configuration file
 	cfg := configparser.ParseConfig()
 
+	// db setup
+	var db *sql.DB
+	if _, err := os.Stat("easyeasm.db"); err == nil {
+		db, _ := sql.Open("sqlite3", "./easyeasm.db")
+		defer db.Close()
+	} else {
+		setupDB()
+		db, _ := sql.Open("sqlite3", "./easyeasm.db")
+		defer db.Close()
+	}
 	// check for previous run file
 	var prevRun bool
 	if _, err := os.Stat("EasyEASM.csv"); err == nil {
@@ -31,11 +44,10 @@ func main() {
 		if e != nil {
 			panic(e)
 		}
-		var domains = utils.getActiveDomains("./easyeasm.db")
+		var domains = getActiveDomains(db)
 		fmt.Println("Active domains from previous run: ", len(domains))
 	} else {
 		fmt.Println("No previous run data found")
-		utils.setupDB()
 		prevRun = false
 	}
 
@@ -58,6 +70,11 @@ func main() {
 		fmt.Printf("\x1b[31mFound %d subdomains\n\n\x1b[0m", Runner.Results)
 		fmt.Println(Runner.Subdomains)
 		fmt.Println("Checking which domains are live and generating assets csv...")
+		if !prevRun {
+			for _, domain := range Runner.Subdomains {
+				insertDomain(db, domain)
+			}
+		}
 
 		// run Httpx to check live domains
 		Runner.RunHttpx()
@@ -66,11 +83,9 @@ func main() {
 		if prevRun && strings.Contains(cfg.RunConfig.SlackWebhook, "https") {
 			utils.NotifyNewDomainsSlack(Runner.Subdomains, cfg.RunConfig.SlackWebhook)
 			os.Remove("old_EasyEASM.csv")
-			os.Remove("old_tracked.csv")
 		} else if prevRun && strings.Contains(cfg.RunConfig.DiscordWebhook, "https") {
 			utils.NotifyNewDomainsDiscord(Runner.Subdomains, cfg.RunConfig.DiscordWebhook)
 			os.Remove("old_EasyEASM.csv")
-			os.Remove("old_tracked.csv")
 		}
 	} else if strings.ToLower(cfg.RunConfig.RunType) == "complete" {
 		// complete run: passive and active enumeration
@@ -103,11 +118,6 @@ func main() {
 		// httpx scan
 		fmt.Printf("Found %d subdomains: ", ActiveRunner.Results)
 		fmt.Println(ActiveRunner.Subdomains)
-		if !prevRun {
-			for _, domain := range ActiveRunner.Subdomains {
-				utils.insertDomain(domain, "./easyeasm.db")
-			}
-		}
 		fmt.Println("Checking which domains are live and generating assets csv...")
 		ActiveRunner.RunHttpx()
 
@@ -115,14 +125,112 @@ func main() {
 		if prevRun && strings.Contains(cfg.RunConfig.SlackWebhook, "https") {
 			utils.NotifyNewDomainsSlack(ActiveRunner.Subdomains, cfg.RunConfig.SlackWebhook)
 			os.Remove("old_EasyEASM.csv")
-			os.Remove("old_tracked.csv")
 		} else if prevRun && strings.Contains(cfg.RunConfig.DiscordWebhook, "https") {
 			utils.NotifyNewDomainsDiscord(ActiveRunner.Subdomains, cfg.RunConfig.DiscordWebhook)
 			os.Remove("old_EasyEASM.csv")
-			os.Remove("old_tracked.csv")
 		}
 	} else {
 		// invalid run mode specified
 		panic("Please pick a valid run mode and add it to your config.yml file! You can set runType to either 'fast' or 'complete'")
+	}
+}
+
+func setupDB() {
+	file, err := os.Create("easyeasm.db")
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+	file.Close()
+
+	db, _ := sql.Open("sqlite3", "./easyeasm.db") // Open the created SQLite File
+	defer db.Close()                              // Defer Closing the database
+	createTable(db)                               // Create Database Tables
+}
+
+func createTable(db *sql.DB) {
+	createDomainTable := `CREATE TABLE domains (
+		"id" integer NOT NULL PRIMARY KEY AUTOINCREMENT,		
+		"domain" TEXT,
+		"active" BOOLEAN,
+		"live" BOOLEAN,
+		"first_seen" TEXT,
+		"last_seen" TEXT		
+	  );`
+
+	statement, err := db.Prepare(createDomainTable)
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+	statement.Exec()
+}
+
+func getDomains(db *sql.DB) []string {
+	rows, err := db.Query("SELECT domain FROM domains")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer rows.Close()
+	var domains []string
+	for rows.Next() {
+		var domain string
+		rows.Scan(&domain)
+		domains = append(domains, domain)
+	}
+	return domains
+}
+
+func getActiveDomains(db *sql.DB) []string {
+	rows, err := db.Query("SELECT domain FROM domains WHERE active = 1")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer rows.Close()
+	var domains []string
+	for rows.Next() {
+		var domain string
+		rows.Scan(&domain)
+		domains = append(domains, domain)
+	}
+	return domains
+}
+
+func getLiveDomains(db *sql.DB) []string {
+	rows, err := db.Query("SELECT domain FROM domains WHERE live = 1")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer rows.Close()
+	var domains []string
+	for rows.Next() {
+		var domain string
+		rows.Scan(&domain)
+		domains = append(domains, domain)
+	}
+	return domains
+}
+
+func insertDomain(db *sql.DB, domain string) {
+	insertSQL := `INSERT INTO domains(domain, active, live, first_seen, last_seen) VALUES (?, ?, ?, ?, ?)`
+	statement, err := db.Prepare(insertSQL)
+	if err != nil {
+		log.Fatalln(err.Error())
+	}
+	var now = time.Now()
+	_, err = statement.Exec(domain, 0, 1, now, now)
+	if err != nil {
+		log.Fatalln(err.Error())
+	}
+}
+
+func updateDomain(db *sql.DB, domain string, live bool) {
+	updateSQL := `UPDATE domains SET live = ?, last_seen = ? WHERE domain = ?`
+	statement, err := db.Prepare(updateSQL)
+	if err != nil {
+		log.Fatalln(err.Error())
+	}
+	var now = time.Now()
+	_, err = statement.Exec(live, now, domain)
+	if err != nil {
+		log.Fatalln(err.Error())
 	}
 }
