+++ pkg/utils/utils.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"bytes"
-	"database/sql"
 	"encoding/csv"
 	"encoding/json"
 	"fmt"
@@ -10,7 +9,6 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"time"
 
 	_ "github.com/mattn/go-sqlite3"
 )
@@ -235,105 +233,5 @@ func installGoTool(name string, path string) {
 		log.Fatalf("An error occurred while installing the package: %s\n%s", err, cmdOutput)
 	}
 
-	log.Printf("Successfully installed the package: %s", packagePath)
-}
-
-func setupDB() {
-	file, err := os.Create("easyeasm.db")
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	file.Close()
-
-	sqliteDatabase, _ := sql.Open("sqlite3", "./sqlite-database.db") // Open the created SQLite File
-	defer sqliteDatabase.Close()                                     // Defer Closing the database
-	createTable(sqliteDatabase)                                      // Create Database Tables
-}
-
-func createTable(db *sql.DB) {
-	createDomainTable := `CREATE TABLE domains (
-		"id" integer NOT NULL PRIMARY KEY AUTOINCREMENT,		
-		"domain" TEXT,
-		"active" BOOLEAN,
-		"live" BOOLEAN,
-		"first_seen" TEXT,
-		"last_seen" TEXT		
-	  );`
-
-	statement, err := db.Prepare(createDomainTable)
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-	statement.Exec()
-}
-
-func getDomains(db *sql.DB) []string {
-	rows, err := db.Query("SELECT domain FROM domains")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rows.Close()
-	var domains []string
-	for rows.Next() {
-		var domain string
-		rows.Scan(&domain)
-		domains = append(domains, domain)
-	}
-	return domains
-}
-
-func getActiveDomains(db *sql.DB) []string {
-	rows, err := db.Query("SELECT domain FROM domains WHERE active = 1")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rows.Close()
-	var domains []string
-	for rows.Next() {
-		var domain string
-		rows.Scan(&domain)
-		domains = append(domains, domain)
-	}
-	return domains
-}
-
-func getLiveDomains(db *sql.DB) []string {
-	rows, err := db.Query("SELECT domain FROM domains WHERE live = 1")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rows.Close()
-	var domains []string
-	for rows.Next() {
-		var domain string
-		rows.Scan(&domain)
-		domains = append(domains, domain)
-	}
-	return domains
-}
-
-func insertDomain(db *sql.DB, domain string) {
-	insertSQL := `INSERT INTO domains(domain, active, live, first_seen, last_seen) VALUES (?, ?, ?, ?, ?)`
-	statement, err := db.Prepare(insertSQL)
-	if err != nil {
-		log.Fatalln(err.Error())
-	}
-	var now = time.Now()
-	_, err = statement.Exec(domain, 0, 1, now, now)
-	if err != nil {
-		log.Fatalln(err.Error())
-	}
-}
-
-func updateDomain(db *sql.DB, domain string, live bool) {
-	updateSQL := `UPDATE domains SET live = ?, last_seen = ? WHERE domain = ?`
-	statement, err := db.Prepare(updateSQL)
-	if err != nil {
-		log.Fatalln(err.Error())
-	}
-	var now = time.Now()
-	_, err = statement.Exec(live, now, domain)
-	if err != nil {
-		log.Fatalln(err.Error())
-	}
+	log.Printf("Successfully installed the %s package: %s", name, packagePath)
 }
